+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.13.0
 
-* Updates platform view creation parameters to use Pigeon.
+* Updates map configuration and platform view creation parameters to use Pigeon.
 
 ## 2.12.0