+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.11.0
 
-* Updates platform view creation parameters to use Pigeon.
+* Updates map configuration and platform view creation parameters to use Pigeon.
 
 ## 2.10.0